requirement.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Mail==0.9.1
mongoengine==0.18.2
Pillow==6.2.1
blinker==1.4
Click==7.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pymongo==3.9.0
six==1.13.0
Werkzeug==0.16.0
